class-42-reactjs-django/demos/budget_tool/requirements.txt
autopep8==1.4.3
confusable-homoglyphs==3.2.0
Django==2.2
django-cors-headers==2.5.2
django-registration==3.0
django-sass-processor==0.7.3
djangorestframework==3.9.2
factory-boy==2.11.1
Faker==1.0.4
libsass==0.18.0
pep8==1.7.1
psycopg2-binary==2.8.1
pycodestyle==2.5.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
